v0.8: improve jittering
In the previous version, a jitter value was calculated for every pixel.
it was then multiplied by a precomputed quasi-random number, to give a
set of per-pixel-unique numbers to generate normal samples.

The thing is, multiplying the two numbers together gave an extremely
skewed distribution, ruining the quality of the reflections (specially
noticeable at higher sample counts)

We now add the two numbers instead, and the jitter gets computed once
per pixel instead of once per ray.

Also, I got rid of a precomputed table of blue noise that was not being
used anymore.

Finally, I touched up some comments.